✨ Added CI workflow for automatic deployment
This commit adds a new file `.github/workflows/ci.yml` which contains the configuration for continuous integration. The workflow is triggered on push events to the repository. It includes a job named "deploy" that runs on Ubuntu latest version. The job uses Node.js version 16.x and performs the following steps:
- Checks out the code using `actions/checkout@v3`
- Sets up Node.js environment using `actions/setup-node@v3`
- Pushes changes to the server using `appleboy/ssh-action@master`
- Executes commands on the server including pulling latest changes, installing dependencies with yarn, building the project, and restarting PM2 process.

The CI workflow enables automated deployment of changes to the specified server based on pushes to the 'rim' branch.